# docs/openapi.yaml
basePath: /api/v1
definitions:
  apierror.Response:
    properties:
      code:
        type: string
        description: Stable machine-readable error code. Clients should branch on this field rather than
          message text.
      details:
        additionalProperties:
          type: string
        type: object
        description: Optional field-level validation details keyed by request field.
      message:
        type: string
        description: Safe human-readable error summary.
      request_id:
        type: string
        description: Server-generated ID for tracing this request; also returned in X-Request-ID.
    type: object
    description: Response representation used by this API.
    example:
      code: INVALID_REQUEST
      message: Invalid request
      request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
      details:
        total: must be a decimal string
  auth.ExternalTokenResponse:
    properties:
      expires_in:
        type: integer
      id:
        type: string
        description: Identifier of the associated resource.
      token:
        type: string
    type: object
  auth.MeResponse:
    properties:
      csrf_token:
        type: string
      profile:
        $ref: '#/definitions/model.UserDto'
    type: object
  auth.externalTokenRequest:
    properties:
      user_id:
        type: string
        description: Identifier of the associated resource.
    required:
    - user_id
    type: object
...
  model.UserDto:
    properties:
      created_at:
        type: string
        format: date-time
        description: Timestamp in RFC 3339 format.
      email:
        type: string
      group_id:
        type: string
        description: Current group ID, or null when the account has no group.
      id:
        type: string
        description: Identifier of the associated resource.
      name:
        type: string
      nick_name:
        type: string
      remaining_coin:
        example: '888.88'
        type: string
        description: Account balance as a JSON string with two fractional digits.
      role_id:
        type: string
        description: Identifier of the associated resource.
    type: object
    description: UserDto representation used by this API.
...
host: localhost:8080
info:
  title: Intania 888 API
  version: 1.0.0-breaking
  description: "Browser and application API for Intania 888. All application routes use the /api/v1 base\
    \ path. Browser authentication uses an HttpOnly session cookie; call GET /auth/me after login to retrieve\
    \ the current profile and session-bound csrf_token. Send that token in X-CSRF-Token with browser mutations\
    \ while a session is active and send an Origin allowed by the server. Mutations are rejected when\
    \ either check fails. \n\nMoney values are JSON strings. Responses use exactly two fractional digits;\
    \ requests accept up to two. Rate values are JSON strings with up to six fractional digits in requests\
    \ and six in responses. Numeric JSON tokens are not accepted for these decimal values. \n\nAPI failures\
    \ use the shared apierror.Response shape: code, message, request_id, and optional field-level details.\
    \ The same request ID is returned in the X-Request-ID header. Dates and timestamps use the formats\
    \ documented by each field. The separate /external/* backend integration is described in docs/README.md\
    \ and is intentionally outside this browser-focused reference.\n\nThe API is rate limited to 200 requests\
    \ per minute per client IP. Exceeding the limit returns 429 TOO_MANY_REQUESTS."
paths:
  /auth/callback:
    get:
      description: Google redirects the browser here with code and state. The handler checks the state
        against the short-lived OAuth cookie, establishes the browser session, clears the OAuth cookie,
        and redirects to the configured frontend. This is a browser redirect; the response is not JSON.
      parameters:
      - description: OAuth authorization code
        in: query
        name: code
        required: true
        type: string
      - description: OAuth state nonce
        in: query
        name: state
        required: true
        type: string
      produces:
      - application/json
      responses:
        '302':
          description: fixed frontend redirect
          schema:
            type: string
          headers:
            X-Request-ID:
              type: string
              description: Correlation ID for this request. The same value appears in apierror.Response.request_id.
        '400':
          description: 'invalid OAuth request Failure body follows apierror.Response; code is stable,
            message is safe to display, and request_id matches the X-Request-ID response header. Failure
            body follows apierror.Response: stable code, safe message, request_id, and optional field
            details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: &id001
              type: string
              description: Correlation ID for this request. The same value appears in apierror.Response.request_id.
          examples:
            application/json:
              code: INVALID_REQUEST
              message: Invalid request
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '403':
          description: 'email is not allowed Failure body follows apierror.Response; code is stable, message
            is safe to display, and request_id matches the X-Request-ID response header. Failure body
            follows apierror.Response: stable code, safe message, request_id, and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: FORBIDDEN
              message: The request is not allowed
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '500':
          description: 'post-login redirect is not configured Failure body follows apierror.Response;
            code is stable, message is safe to display, and request_id matches the X-Request-ID response
            header. Failure body follows apierror.Response: stable code, safe message, request_id, and
            optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: INTERNAL_ERROR
              message: Internal server error
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '503':
          description: 'OAuth login or access policy unavailable Failure body follows apierror.Response;
            code is stable, message is safe to display, and request_id matches the X-Request-ID response
            header. Failure body follows apierror.Response: stable code, safe message, request_id, and
            optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: DEPENDENCY_UNAVAILABLE
              message: A required dependency is unavailable
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '429':
          description: The client IP exceeded the API limit of 200 requests per minute.
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID:
              type: string
              description: Correlation ID for this request. The same value appears in apierror.Response.request_id.
          examples:
            application/json:
              code: TOO_MANY_REQUESTS
              message: Too many requests
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
      summary: Complete Google OAuth login
      tags:
      - Auth
  /auth/external-tokens:
    post:
      parameters:
      - description: existing user
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.externalTokenRequest'
      - name: X-CSRF-Token
        in: header
        description: Session-bound token returned by GET /auth/me. Required for browser mutations.
        required: true
        type: string
      responses:
        '201':
          description: Created
          schema:
            $ref: '#/definitions/auth.ExternalTokenResponse'
          headers:
            X-Request-ID:
              type: string
              description: Correlation ID for this request. The same value appears in apierror.Response.request_id.
        '400':
          description: 'Bad Request Failure body follows apierror.Response; code is stable, message is
            safe to display, and request_id matches the X-Request-ID response header. Failure body follows
            apierror.Response: stable code, safe message, request_id, and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: INVALID_REQUEST
              message: Invalid request
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '401':
          description: 'Unauthorized Failure body follows apierror.Response; code is stable, message is
            safe to display, and request_id matches the X-Request-ID response header. Failure body follows
            apierror.Response: stable code, safe message, request_id, and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: UNAUTHORIZED
              message: Authentication required
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '403':
          description: 'Forbidden Failure body follows apierror.Response; code is stable, message is safe
            to display, and request_id matches the X-Request-ID response header. Failure body follows
            apierror.Response: stable code, safe message, request_id, and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: FORBIDDEN
              message: The request is not allowed
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '404':
          description: 'Not Found Failure body follows apierror.Response; code is stable, message is safe
            to display, and request_id matches the X-Request-ID response header. Failure body follows
            apierror.Response: stable code, safe message, request_id, and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
        '503':
          description: 'Service Unavailable Failure body follows apierror.Response; code is stable, message
            is safe to display, and request_id matches the X-Request-ID response header. Failure body
            follows apierror.Response: stable code, safe message, request_id, and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: DEPENDENCY_UNAVAILABLE
              message: A required dependency is unavailable
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '429':
          description: The client IP exceeded the API limit of 200 requests per minute.
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID:
              type: string
              description: Correlation ID for this request. The same value appears in apierror.Response.request_id.
          examples:
            application/json:
              code: TOO_MANY_REQUESTS
              message: Too many requests
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
      security:
      - CookieSession: []
      summary: Issue a one-hour external JWT
      tags:
      - Auth
      description: Deprecated. Administrator-only. Issues a one-hour Bearer JWT for the selected account.
        The response includes the token, its revocation ID, and its lifetime in seconds. Treat the token
        as a secret; use DELETE on the returned ID to revoke it.
      deprecated: true
  /auth/external-tokens/{id}:
    delete:
      parameters:
      - description: token ID
        in: path
        name: id
        required: true
        type: string
      - name: X-CSRF-Token
        in: header
        description: Session-bound token returned by GET /auth/me. Required for browser mutations.
        required: true
        type: string
      responses:
        '204':
          description: revoked
          headers:
            X-Request-ID:
              type: string
              description: Correlation ID for this request. The same value appears in apierror.Response.request_id.
        '401':
          description: 'Unauthorized Failure body follows apierror.Response; code is stable, message is
            safe to display, and request_id matches the X-Request-ID response header. Failure body follows
            apierror.Response: stable code, safe message, request_id, and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: UNAUTHORIZED
              message: Authentication required
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '403':
          description: 'Forbidden Failure body follows apierror.Response; code is stable, message is safe
            to display, and request_id matches the X-Request-ID response header. Failure body follows
            apierror.Response: stable code, safe message, request_id, and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: FORBIDDEN
              message: The request is not allowed
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '503':
          description: 'Service Unavailable Failure body follows apierror.Response; code is stable, message
            is safe to display, and request_id matches the X-Request-ID response header. Failure body
            follows apierror.Response: stable code, safe message, request_id, and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: DEPENDENCY_UNAVAILABLE
              message: A required dependency is unavailable
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '429':
          description: The client IP exceeded the API limit of 200 requests per minute.
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID:
              type: string
              description: Correlation ID for this request. The same value appears in apierror.Response.request_id.
          examples:
            application/json:
              code: TOO_MANY_REQUESTS
              message: Too many requests
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
      security:
      - CookieSession: []
      summary: Revoke an external JWT
      tags:
      - Auth
      description: Deprecated. Administrator-only. Revokes an issued external JWT by its revocation ID.
        The operation returns no response body after successful revocation.
      deprecated: true
  /auth/login:
    get:
      description: Starts Google OAuth and returns the authorization URL to open in the browser. It also
        sets a short-lived HttpOnly state cookie that must return with the OAuth callback. The server
        controls the post-login destination; redirect_to is not supported.
      produces:
      - application/json
      responses:
        '200':
          description: authorization URL
          schema:
            additionalProperties:
              type: string
            type: object
          headers:
            X-Request-ID:
              type: string
              description: Correlation ID for this request. The same value appears in apierror.Response.request_id.
        '400':
          description: 'redirect_to is not supported Failure body follows apierror.Response; code is stable,
            message is safe to display, and request_id matches the X-Request-ID response header. Failure
            body follows apierror.Response: stable code, safe message, request_id, and optional field
            details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: INVALID_REQUEST
              message: Invalid request
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '503':
          description: 'OAuth login unavailable Failure body follows apierror.Response; code is stable,
            message is safe to display, and request_id matches the X-Request-ID response header. Failure
            body follows apierror.Response: stable code, safe message, request_id, and optional field
            details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: DEPENDENCY_UNAVAILABLE
              message: A required dependency is unavailable
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '429':
          description: The client IP exceeded the API limit of 200 requests per minute.
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID:
              type: string
              description: Correlation ID for this request. The same value appears in apierror.Response.request_id.
          examples:
            application/json:
              code: TOO_MANY_REQUESTS
              message: Too many requests
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
      summary: Start Google OAuth login
      tags:
      - Auth
  /auth/logout:
    post:
      description: Revokes the active server-side browser session and expires its cookie. If there is
        no active session, logout succeeds idempotently. When a session is active, send the session-bound
        X-CSRF-Token from GET /auth/me.
      responses:
        '204':
          description: logged out
          headers:
            X-Request-ID:
              type: string
              description: Correlation ID for this request. The same value appears in apierror.Response.request_id.
        '403':
          description: 'invalid CSRF token Failure body follows apierror.Response; code is stable, message
            is safe to display, and request_id matches the X-Request-ID response header. Failure body
            follows apierror.Response: stable code, safe message, request_id, and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: FORBIDDEN
              message: The request is not allowed
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '503':
          description: 'session store or revocation unavailable Failure body follows apierror.Response;
            code is stable, message is safe to display, and request_id matches the X-Request-ID response
            header. Failure body follows apierror.Response: stable code, safe message, request_id, and
            optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: DEPENDENCY_UNAVAILABLE
              message: A required dependency is unavailable
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '429':
          description: The client IP exceeded the API limit of 200 requests per minute.
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID:
              type: string
              description: Correlation ID for this request. The same value appears in apierror.Response.request_id.
          examples:
            application/json:
              code: TOO_MANY_REQUESTS
              message: Too many requests
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
      security:
      - CookieSession: []
      summary: Log out of browser session
      tags:
      - Auth
      parameters:
      - name: X-CSRF-Token
        in: header
        description: Session-bound token returned by GET /auth/me. Required when a session is active;
          logout succeeds without a session.
        required: false
        type: string
  /auth/me:
    get:
      description: Returns the profile for the current browser session and a session-bound csrf_token.
        Keep the token in memory and send it in X-CSRF-Token on browser mutations; do not persist it in
        local storage.
      produces:
      - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/auth.MeResponse'
          headers:
            X-Request-ID:
              type: string
              description: Correlation ID for this request. The same value appears in apierror.Response.request_id.
          examples:
            application/json:
              csrf_token: a-session-bound-token
              profile:
                id: 9c90dcf6-09f7-4e46-a06d-8dbfeec2b981
                name: Example User
                nick_name: Alex
                group_id: blue
        '401':
          description: 'unauthorized Failure body follows apierror.Response; code is stable, message is
            safe to display, and request_id matches the X-Request-ID response header. Failure body follows
            apierror.Response: stable code, safe message, request_id, and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: UNAUTHORIZED
              message: Authentication required
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '503':
          description: 'session or policy unavailable Failure body follows apierror.Response; code is
            stable, message is safe to display, and request_id matches the X-Request-ID response header.
            Failure body follows apierror.Response: stable code, safe message, request_id, and optional
            field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: DEPENDENCY_UNAVAILABLE
              message: A required dependency is unavailable
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '429':
          description: The client IP exceeded the API limit of 200 requests per minute.
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID:
              type: string
              description: Correlation ID for this request. The same value appears in apierror.Response.request_id.
          examples:
            application/json:
              code: TOO_MANY_REQUESTS
              message: Too many requests
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
      security:
      - CookieSession: []
      summary: Browser profile and CSRF token
      tags:
      - Auth
...
schemes:
- http
securityDefinitions:
  CookieSession:
    description: Browser session cookie. The cookie is HttpOnly and is normally sent by the browser automatically.
      Production uses the __Host-session cookie name. Swagger 2.0 represents this as a Cookie header.
      For browser mutations, also send X-CSRF-Token from GET /auth/me and an allowed Origin.
    in: header
    name: Cookie
    type: apiKey
swagger: '2.0'
consumes:
- application/json
produces:
- application/json
...